pyproject.toml
[project]
name = "powercontroller"
version = "1.2.3"
description = "App to control and monitor devices via Shelly smart switches"
readme = "README.md"
requires-python = ">=3.13"
service_name="PowerController"
launch_path="src/main.py"
dependencies = [
    "astral>=3.2",
    "fastapi>=0.115.0",
    "flask>=3.1.2",
    "jinja2>=3.1.0",
    "org-enums>=1.0.0",
    "pre-commit>=4.3.0",
    "python-dateutil>=2.9.0.post0",
    "pytz>=2025.2",
    "sc-utility>=2.4.12",
    "uvicorn[standard]>=0.32.0",
    "werkzeug>=3.1.3",
    "psycopg[binary]>=3.2",
    "python-dotenv>=1.0.1",
]

[tool.uv.sources]
org-enums = { git = "https://github.com/NickElseySpelloC/org_enums", tag = "v1.0.0" }

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

